refactor: extract CommunitySelectCard and formatCompactNumber from SelectAudienceModal
Split SelectAudienceModal into separate files for better separation of concerns:
- Extract CommunitySelectCard component into its own file
- Move formatNumber utility to lib/utils as formatCompactNumber
- Update barrel exports in shared/index.ts
