+++ src/components/shared/index.ts
@@ -1,5 +1,8 @@
 export { Modal } from './Modal'
 export type { ModalProps } from './Modal'
 
+export { CommunitySelectCard } from './CommunitySelectCard'
+export type { CommunitySelectCardProps } from './CommunitySelectCard'
+
 export { SelectAudienceModal } from './SelectAudienceModal'
 export type { SelectAudienceModalProps } from './SelectAudienceModal'
