+++ src/lib/utils.ts
@@ -21,6 +21,16 @@ export function formatPercentage(value: number): string {
   return `${sign}${value.toFixed(1)}%`
 }
 
+export function formatCompactNumber(num: number): string {
+  if (num >= 1_000_000) {
+    return `${(num / 1_000_000).toFixed(1)}M`
+  }
+  if (num >= 1_000) {
+    return `${(num / 1_000).toFixed(0)}K`
+  }
+  return num.toString()
+}
+
 export function formatDate(date: Date | string): string {
   const d = typeof date === 'string' ? new Date(date) : date
   return new Intl.DateTimeFormat('en-US', {
